pyproject.toml
[build-system]
requires = [
    "setuptools",
    "autocalver",
]
build-backend = "setuptools.build_meta"

[project]
name = "silly_pyproject_name"
dynamic = ["version"]
description = "CLI tools for manipulating Python project's name. Example for Gather."
readme = "README.rst"
authors = [{name = "Moshe Zadka", email = "moshez@zadka.club"}]
dependencies = ["gather", "tomlkit"]

[project.optional-dependencies]
tests = ["virtue", "pyhamcrest", "coverage"]
mypy = ["mypy"]
lint = ["flake8", "black"]
docs = ["sphinx"]

[project.license]
text = """
Permission is hereby granted, free of charge, to any person obtaining a
copy of this software and associated documentation files (the "Software"),
to deal in the Software without restriction, including without limitation the
rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is furnished
to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
"""

[project.urls]
Homepage = "https://github.com/elcaminoreal/silly_pyproject_name"

[tool.autocalver]
use = true
log = "git-log-head"
log_command = "git log -n 1 --date=iso"
is_main_var = "GITHUB_REF"
is_main_match = ".*/trunk$"

[project.entry-points.gather]
placeholder = "silly_pyproject_name"

# This allows running the scripts using
# script entry points. Note that `python -m silly_project_name` still
# works correctly.
[project.scripts]
# The eponymous command allows running all subcommands.
silly-pyproject-name = "silly_pyproject_name:ENTRY_DATA.main_command"
# This allows running a subcommand directly.
silly-pyproject-name-rename = "silly_pyproject_name:ENTRY_DATA.sub_command"

